Add hot reload support

+++ src/conf/types/runtime/client_lower.rs
@@ -4,8 +4,8 @@ use crate::conf::types::specification::client::ClientSpec;
 impl From<ClientSpec> for ClientConfig {
     fn from(spec: ClientSpec) -> Self {
         ClientConfig {
-            read_buffer_bytes: spec.read_buffer_bytes as usize,
-            pub_sub_channel_capacity: spec.pub_sub_channel_capacity as usize,
+            read_buffer_bytes: spec.read_buffer_bytes,
+            pub_sub_channel_capacity: spec.pub_sub_channel_capacity,
         }
     }
 }
